Workaround for wiki filenames with non-ascii chars
Since "\" symbols can't be in filenames, the caret (^) is used instead.

Replace all instances of carets with a backslash and then to make the text
contain non-ascii characters, feed the string through like this:

echo -en "This is a sp\xe9cial string"

The above line outputs (without a newline): This is a spécial string.